pyproject.toml
[tool.poetry]
name = "home-work-drf"
version = "0.1.0"
description = ""
authors = ["Lewbertsklol <lewbertsklol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
django-types = "^0.19.1"
drf-yasg = "^1.21.7"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
